Add entry function for tracker
Currently, the tracker does not allow users to track input.

A add entry function is created to allow users to track their incomes as
well as spendings. The function call is ivoked with 'add' keyword. The
format is of add[add type][amount][description]. Add type can be of
'income' or 'spending' types right now.