Add login feature to the bot
Currently, the bot is unable to secure the information of the user and
with income as one of the storage for the bot, there needs to have a
login system to secure these information for the user.

The bot implements a simple password verification at the start to ensure
the user is the one using the bot.